Move presets XML to user folder
Change location of XML file containing saved presets to
~/.config/continue-folder/

Previously, the XML file was saved in the same folder as the python
script.  This is fine until a studio wants to deploy this script in a
central location, then every artist will be accessing the same presets
XML.  The presets are meant to be per artist, so move to a user specific
folder.